refactore: split programm to modules
